chore: update command position
